+++ .circleci/setupAndTestWithAuthReact.sh
@@ -57,6 +57,11 @@ cd ../../
 cd ../project/test/auth-react-server
 go run main.go &
 cd ../../../supertokens-auth-react/
+
+# When testing with supertokens-auth-react for version >= 0.18 the SKIP_OAUTH 
+# flag will not be checked because Auth0 is used as a provider so that the Thirdparty tests can run reliably. 
+# In versions lower than 0.18 Github is used as the provider.
+
 SKIP_OAUTH=true npm run test-with-non-node
 if [[ $? -ne 0 ]]
 then