+++ recipe/openid/openidmodels/apiInterface.go
@@ -13,8 +13,19 @@
  * under the License.
  */
 
-package jwt
+package openidmodels
 
-const (
-	GetJWKSAPI = "/jwt/jwks.json"
-)
+import "net/http"
+
+type APIOptions struct {
+	RecipeImplementation RecipeInterface
+	Config               TypeNormalisedInput
+	RecipeID             string
+	Req                  *http.Request
+	Res                  http.ResponseWriter
+	OtherHandler         http.HandlerFunc
+}
+
+type APIInterface struct {
+	GetOpenIdDiscoveryConfigurationGET *func(options APIOptions) (GetOpenIdDiscoveryConfigurationResponse, error)
+}