+++ recipe/openid/openidmodels/models.go
@@ -13,8 +13,28 @@
  * under the License.
  */
 
-package jwt
+package openidmodels
 
-const (
-	GetJWKSAPI = "/jwt/jwks.json"
+import (
+	"github.com/supertokens/supertokens-golang/recipe/jwt/jwtmodels"
+	"github.com/supertokens/supertokens-golang/supertokens"
 )
+
+type TypeInput struct {
+	Issuer             *string
+	JwtValiditySeconds *uint64
+	Override           *OverrideStruct
+}
+
+type TypeNormalisedInput struct {
+	IssuerDomain       supertokens.NormalisedURLDomain
+	IssuerPath         supertokens.NormalisedURLPath
+	JwtValiditySeconds *uint64
+	Override           OverrideStruct
+}
+
+type OverrideStruct struct {
+	Functions  func(originalImplementation RecipeInterface) RecipeInterface
+	APIs       func(originalImplementation APIInterface) APIInterface
+	JwtFeature *jwtmodels.OverrideStruct
+}