+++ supertokens/utils.go
@@ -137,8 +137,8 @@ func getRIDFromRequest(r *http.Request) string {
 }
 
 func Send200Response(res http.ResponseWriter, responseJson interface{}) error {
-	dw := res.(*DoneWriter)
-	if !dw.done {
+	dw := MakeDoneWriter(res)
+	if !dw.IsDone() {
 		res.Header().Set("Content-Type", "application/json; charset=utf-8")
 		res.WriteHeader(200)
 		bytes, err := json.Marshal(responseJson)
@@ -152,12 +152,12 @@ func Send200Response(res http.ResponseWriter, responseJson interface{}) error {
 }
 
 func SendNon200Response(res http.ResponseWriter, message string, statusCode int) error {
-	dw := res.(*DoneWriter)
-	if !dw.done {
+	dw := MakeDoneWriter(res)
+	if !dw.IsDone() {
 		if statusCode < 300 {
-			return errors.New("Calling sendNon200Response with status code < 300")
+			return errors.New("calling sendNon200Response with status code < 300")
 		}
-		res.Header().Set("Content-Type", "text/html; charset=utf-8")
+		res.Header().Set("Content-Type", "application/json; charset=utf-8")
 		res.WriteHeader(statusCode)
 		response := map[string]interface{}{
 			"message": message,